% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2b_sol.R
\name{p2b_sol}
\alias{p2b_sol}
\title{Function to calculate distribution of sample means.}
\usage{
p2b_sol(n, x, lambda = 10)
}
\arguments{
\item{n}{The size of each sample from which to calculate a mean}

\item{x}{The number of samples to calculate.}

\item{lambda}{The lone parameter of the poisson distribution.}
}
\description{
Function to calculate distribution of sample means.
}
